+++ core/tests/test_helpers.py
@@ -1,7 +1,8 @@
 from unittest import mock
 
-from core import helpers
+import pytest
 
+from core import helpers
 from core.tests.helpers import create_response
 
 
@@ -31,9 +32,11 @@ def test_create_user_profile(mock_create_user_profile, mock_profile_update):
     )
 
 
+@pytest.mark.parametrize('status_code', [200, 404])
 @mock.patch.object(helpers.api_client.supplier, 'profile_update')
 @mock.patch.object(helpers.sso_api_client.user, 'update_user_profile')
-def test_update_user_profile(mock_update_user_profile, mock_profile_update):
+def test_update_user_profile(mock_update_user_profile, mock_profile_update, status_code):
+    mock_profile_update.return_value = create_response(status_code=status_code)
     data = {
         'first_name': 'First Name',
         'last_name': 'Last Name',
@@ -43,16 +46,9 @@ def test_update_user_profile(mock_update_user_profile, mock_profile_update):
     profile_name_data = {'name': data['first_name'] + ' ' + data['last_name']}
 
     mock_update_user_profile.return_value = create_response(status_code=201, json_body=data)
-    helpers.update_user_profile(
-        sso_session_id=1,
-        data=data
-    )
-    assert mock_update_user_profile.call_count == 1
-    assert mock_update_user_profile.call_args == mock.call(
-        sso_session_id=1, data=data
-    )
+    helpers.update_user_profile(sso_session_id=1, data=data)
 
+    assert mock_update_user_profile.call_count == 1
+    assert mock_update_user_profile.call_args == mock.call(sso_session_id=1, data=data)
     assert mock_profile_update.call_count == 1
-    assert mock_profile_update.call_args == mock.call(
-        sso_session_id=1, data=profile_name_data
-    )
+    assert mock_profile_update.call_args == mock.call(sso_session_id=1, data=profile_name_data)