+++ CHANGELOG.md
@@ -63,6 +63,7 @@
 - TT-1904 - admin defaults to invite collaborators 
 - TT-1901 - show name in collaborators list 
 - TT-1914 - Reinstate missing terms tickbox
+- TT-1915 - Fix update profile error if user has no supplier
 
 ## [2019.08.14_1](https://github.com/uktrade/directory-ui-supplier/releases/tag/2019.08.14_1)
 [Full Changelog](https://github.com/uktrade/directory-ui-supplier/compare/2019.08.14...2019.08.14_1)